Chore #56 - 코드리뷰 반영
