+++ Baggle/Baggle/Core/Models/MeetingDetail.swift
@@ -8,8 +8,8 @@
 import Foundation
 
 /// 본인 포함 참여자 모델
-struct Member: Equatable {
-    let userid: Int // 유저 id
+struct Member: Equatable, Identifiable {
+    var id: Int // 유저 id
     let name: String // 유저 이름
     let profileURL: String // 유저 프로필 이미지 URL
     let isOwner: Bool // 방장 여부
@@ -17,7 +17,7 @@ struct Member: Equatable {
     let certImage: String // 인증 사진, 별도 분리 가능 O
 
     static func == (lhs: Member, rhs: Member) -> Bool {
-        return lhs.userid == rhs.userid
+        return lhs.id == rhs.id
     }
 }
